Calculates displacement between every tag location update and merges displacement data with tag location data (Beta)

+++ app/src/main/java/bachelor/test/locationapp/presenter/positioning/LocationData.kt
@@ -1,6 +1,6 @@
 package bachelor.test.locationapp.presenter.positioning
 
-data class LocationData(var xPos: Double, var yPos: Double, var zPos: Double, val qualityFactor: Int) {
+data class LocationData(var xPos: Float, var yPos: Float, var zPos: Float, val qualityFactor: Int) {
 
     override fun toString(): String {
         return "$xPos, $yPos, $zPos, $qualityFactor"
